DiagnoCrew/requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.1